Move python helper modules into library
