+++ src/stagelib/envglobals.py
